requirements.txt
Django>=5.2,<6.0

djangorestframework>=3.16,<3.20

psycopg2-binary>=2.9

python-dotenv>=1.0

celery>=5.5

redis>=5.0

requests>=2.31
